backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
tortoise-orm==0.20.0
asyncpg==0.29.0
pydantic[email]>=2.7.4
pydantic-settings>=2.3.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
langchain==1.0.0
langgraph==1.0.1
langchain-aws==1.0.0
langchain-core>=1.0.0
langchain-community>=0.4
google-generativeai==0.3.2
groq>=0.11.0
boto3>=1.40.19
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==1.1.0
tavily-python==0.3.0
